parsing: resolve wdir for empty string as curdir
This commit does fix the issue within the parsing resolver
when the `wdir` is passed as an empty string. When it is passed
as such, it interprets it as a current working directory which
is a correct assumption.

This still does not fix the issue with fs.path.join. Original issue
with `fs.path.join` is still there.

```pycon
>>> os.path.join('', 'dvc.yaml')
'dvc.yaml'
>>> fs.path.join('', 'dvc.yaml')
'/dvc.yaml'
```