cli: support external commands available in PATH
This commit adds support for external commands that are available
in one of the directory set in PATH.
The executable should be in format of \'dvc-*\' format, should
have executable permissions and need to exist (duh). The suffix,
after 'dvc-' will be used as a subcommand name, and all tha arguments
after that subcommand will be passed to that external command.