dvc.objects: extend diskcache to handle pickle errors
This commit decouples the UI representation with the exception by:
1. extending diskcache to raise a proper DiskError.
2. handling the error in `main` function where we provide a nice
   error message and hints.

Since we no longer can use DvcException in dvc.objects, this had
to be done.

To reduce duplications and error, now `dvc.objects` will provide
the caching infrastructure instead of depending on the `diskcache`.
This makes sense as `dvc.state` is already in `dvc.objects`,
and since `dvc.data` is anyway going to depend on `dvc.objects`,
`dvc.data.ODBIndex` now uses this extended version of `Cache` too.

While I was at it, the new `Cache` was also extended to default
to a constant protocol instead of `pickle.HIGHEST_PROTOCOL`.

Also added the test for UI message.

dvc.objects.cache: extend Disk to provide a way to add disk_type

This is used in UI messages, and we have already documented
the message. And it's trivial to add the support.

We only needed this for ODBIndex, as it creates a random
uid directory inside .dvc/tmp/index directory.

state/index: translate ValueError with messages that contain 'pickle protocol'