fix: allow create a stage when calling process working dir is not in dvc repo
`dvc.yaml` was harcoded in the `_create_stage` method of class `StageCache`. In some use case, we want to cache the stage but the calling process is not in the dvc repo. Hardcoding `dvc.yaml` raises an error (StagePathOutsideError) in the path verification as the dvc.yaml is out of the repository.
By providing a path we can pass the `stage` path file directly. So, now it is possible to create the cache!