fix(strictyaml): display valid lc consistently
We understand that certain data types may lack a `lc` property, so we utilize the `key()` and `item()` methods of the `lc` property to obtain the line and column details for such types. Even though we still need to navigate to the parent mapping or sequence to access the `lc` property, these methods enable us to handle diverse data types seamlessly and display accurate line and column information.

Fixes #10109

Signed-off-by: hqdncw <hqdncw@gmail.com>