fix(diff): do not show missing cache entries on diff result
The directory entry might show up in "Not in cache" when the second rev is a
workspace revision.

But files inside missing cache entries won't be shown in any of the state,
unless we know the oid of the entries from the index on both side of the
`diff`.

On top of #10844. Closes #7661.